Refactor: Auth 모듈 이름 User 모듈로 변경
