+++ be/src/app.module.ts
@@ -7,7 +7,6 @@ import { ConfigModule, ConfigService } from "@nestjs/config";
 import { ScheduleApiModule } from "./schedule-api/schedule-api.module";
 import { AuthApiModule } from "./auth-api/auth-api.module";
 import { CategoryApiModule } from "./category-api/category-api.module";
-import { CategoryModule } from "./category/category.module";
 import dbConfig from "./config/dbConfig";
 
 @Module({
