+++ be/src/auth-api/auth-api.controller.ts
@@ -1,36 +1,36 @@
 import { Controller, Post, Body, HttpCode, HttpStatus, Delete, Patch } from "@nestjs/common";
-import { AuthService } from "../auth/auth.service";
-import { UserLoginDto } from "../auth/dto/user-login.dto";
-import { CreateUserDto } from "../auth/dto/create-user.dto";
-import { UserModifyDto } from "../auth/dto/user-modify.dto";
+import { UserService } from "../user/user.service";
+import { UserLoginDto } from "../user/dto/user-login.dto";
+import { CreateUserDto } from "../user/dto/create-user.dto";
+import { UserModifyDto } from "../user/dto/user-modify.dto";
 
 @Controller("/api/auth")
 export class AuthApiController {
-  constructor(private authService: AuthService) {}
+  constructor(private userService: UserService) {}
 
   @Post("/register")
   async register(@Body() dto: CreateUserDto): Promise<JSON> {
-    const result = await this.authService.register(dto);
+    const result = await this.userService.register(dto);
     return JSON.parse(result);
   }
 
   @Post("/login")
   @HttpCode(HttpStatus.OK)
   async login(@Body() dto: UserLoginDto): Promise<JSON> {
-    const result = await this.authService.login(dto);
+    const result = await this.userService.login(dto);
     return JSON.parse(result);
   }
 
   @Delete("/delete")
   async deleteAccount(@Body() dto: UserLoginDto) {
-    const result = await this.authService.deleteAccount(dto);
+    const result = await this.userService.deleteAccount(dto);
     return JSON.parse(result);
   }
 
   @Patch("/update")
   @HttpCode(HttpStatus.OK)
   async updateUserInfo(@Body() dto: UserModifyDto): Promise<JSON> {
-    const result = await this.authService.update(dto);
+    const result = await this.userService.update(dto);
     return JSON.parse(result);
   }
 }
