+++ be/src/user/user.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { AuthService } from "./auth.service";
+import { UserService } from "./user.service";
 import { UserEntity } from "./entity/user.entity";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { UserRepository } from "./user.repository";
@@ -8,7 +8,7 @@ import { AuthApiController } from "src/auth-api/auth-api.controller";
 @Module({
   imports: [TypeOrmModule.forFeature([UserEntity])],
   controllers: [AuthApiController],
-  providers: [AuthService, UserRepository],
-  exports: [AuthService],
+  providers: [UserService, UserRepository],
+  exports: [UserService],
 })
-export class AuthModule {}
+export class UserModule {}
